refactor set_random_seed
